backend/api/app/requirements.txt
python-dotenv
flask
flask-cors
google-cloud-firestore